Keep the web-editor tests out of the no-extras test run
test-core-install runs the suite against a bare `pip install . pytest` to
catch core code drifting onto optional-extra packages, excluding web tests by
filename with --ignore-glob='tests/test_web_*.py'. Two tests added in the
previous commit import the web editor from tests/test_weather_onecall.py,
which that glob does not match, so the job hit ModuleNotFoundError on flask.

Moves both to tests/test_web_config.py, where web tests live precisely
because they need the extras. The checks that only read config.html and
dashboard.js as text stay in the core suite - they import nothing, and the
enum-drift guard is worth keeping next to the enum.

Verified by reproducing the job locally with flask and waitress made
unimportable: fails on the same single test before the move, clean after.
